requirements.txt
fastapi
pydantic
beanie
motor
uvicorn
python-dotenv
requests
cloudinary
google-api-python-client
google-auth
google-auth-oauthlib
google-genai
gtts
edge-tts
moviepy
pydub
pillow
transformers
huggingface-hub
diffusers
bcrypt
python-jose
markdown2
beautifulsoup4
wikipedia
python-multipart